requirements.txt
annotated-types==0.7.0
anyio==4.6.2
bcrypt==4.2.0
cffi==1.17.1
click==8.1.7
colorama==0.4.6
cryptography==43.0.1
dnspython==2.7.0
ecdsa==0.19.0
fastapi==0.115.2
h11==0.14.0
idna==3.10
passlib==1.7.4
pyasn1==0.6.1
pycparser==2.22
pydantic==2.9.2
pydantic_core==2.23.4
pymongo==4.10.1
python-jose==3.3.0
rsa==4.9
six==1.16.0
sniffio==1.3.1
starlette==0.39.2
typing_extensions==4.12.2
uvicorn==0.31.1
